# internal/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Account struct {
	ID         uuid.UUID `json:"id"`
	ApiKeyHash string    `json:"api_key_hash"`
	Name       string    `json:"name"`
	// durable prepaid balance; mutated on topup/debit/refund, never derived from a log
	Balance   int64              `json:"balance"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type Campaign struct {
	ID                   uuid.UUID          `json:"id"`
	AccountID            uuid.UUID          `json:"account_id"`
	Text                 string             `json:"text"`
	TotalRecipients      int32              `json:"total_recipients"`
	CostPerMessage       int64              `json:"cost_per_message"`
	TotalCost            int64              `json:"total_cost"`
	Status               string             `json:"status"`
	CreatedAt            pgtype.Timestamptz `json:"created_at"`
	UpdatedAt            pgtype.Timestamptz `json:"updated_at"`
	ExpandedThroughIndex int32              `json:"expanded_through_index"`
}

type Message struct {
	ID           uuid.UUID          `json:"id"`
	AccountID    uuid.UUID          `json:"account_id"`
	CampaignID   *uuid.UUID         `json:"campaign_id"`
	Recipient    string             `json:"recipient"`
	Priority     string             `json:"priority"`
	Cost         int64              `json:"cost"`
	Status       string             `json:"status"`
	Operator     pgtype.Text        `json:"operator"`
	DeadlineAt   pgtype.Timestamptz `json:"deadline_at"`
	DispatchedAt pgtype.Timestamptz `json:"dispatched_at"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type MessageStatusEvent struct {
	ID         uuid.UUID          `json:"id"`
	MessageID  uuid.UUID          `json:"message_id"`
	Status     string             `json:"status"`
	OccurredAt pgtype.Timestamptz `json:"occurred_at"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `json:"updated_at"`
}

type MessagesDefault struct {
	ID           uuid.UUID          `json:"id"`
	AccountID    uuid.UUID          `json:"account_id"`
	CampaignID   *uuid.UUID         `json:"campaign_id"`
	Recipient    string             `json:"recipient"`
	Priority     string             `json:"priority"`
	Cost         int64              `json:"cost"`
	Status       string             `json:"status"`
	Operator     pgtype.Text        `json:"operator"`
	DeadlineAt   pgtype.Timestamptz `json:"deadline_at"`
	DispatchedAt pgtype.Timestamptz `json:"dispatched_at"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type ProcessedEvent struct {
	ConsumerName string             `json:"consumer_name"`
	EventID      string             `json:"event_id"`
	ProcessedAt  pgtype.Timestamptz `json:"processed_at"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}

type TopupIdempotency struct {
	AccountID      uuid.UUID          `json:"account_id"`
	IdempotencyKey string             `json:"idempotency_key"`
	Amount         int64              `json:"amount"`
	DurableBalance int64              `json:"durable_balance"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
}